+++ asdf/resource.py
@@ -7,14 +7,6 @@ from pathlib import Path
 import fnmatch
 import os
 import pkgutil
-import sys
-
-if sys.version_info < (3, 9):
-    import importlib_resources
-else:
-    import importlib.resources as importlib_resources
-
-import asdf
 
 from .util import get_class_name
 
@@ -24,7 +16,6 @@ __all__ = [
     "DirectoryResourceMapping",
     "ResourceManager",
     "JsonschemaResourceMapping",
-    "get_core_resource_mappings",
 ]
 
 
@@ -270,31 +261,3 @@ class JsonschemaResourceMapping(Mapping):
 
     def __repr__(self):
         return "JsonschemaResourceMapping()"
-
-
-def get_core_resource_mappings():
-    """
-    Get the resource mapping instances for the core schemas.
-    This method is registered with the asdf.resource_mappings entry point.
-    """
-    core_schemas_root = importlib_resources.files(asdf)/"schemas"/"stsci.edu"
-    if not core_schemas_root.is_dir():
-        # In an editable install, the schemas can be found in the
-        # asdf-standard submodule.
-        core_schemas_root = Path(__file__).parent.parent/"asdf-standard"/"schemas"/"stsci.edu"
-        if not core_schemas_root.is_dir():
-            raise RuntimeError("Unable to locate core schemas")
-
-    resources_root = importlib_resources.files(asdf)/"resources"
-    if not resources_root.is_dir():
-        # In an editable install, the resources can be found in the
-        # asdf-standard submodule.
-        resources_root = Path(__file__).parent.parent/"asdf-standard"/"resources"
-        if not resources_root.is_dir():
-            raise RuntimeError("Unable to locate core resources")
-
-    return [
-        DirectoryResourceMapping(core_schemas_root, "http://stsci.edu/schemas", recursive=True),
-        DirectoryResourceMapping(resources_root / "asdf-format.org", "asdf://asdf-format.org", recursive=True),
-        JsonschemaResourceMapping(),
-    ]