+++ asdf/tests/test_fits_embed.py
@@ -15,7 +15,7 @@ import asdf
 from asdf import get_config
 from asdf import fits_embed
 from asdf import open as asdf_open
-from asdf.exceptions import AsdfWarning, AsdfConversionWarning
+from asdf.exceptions import AsdfWarning
 
 from .helpers import (
     assert_tree_match,
@@ -343,29 +343,6 @@ def test_bad_input(tmpdir):
 
     with pytest.raises(ValueError):
         asdf_open(text_file)
-
-def test_version_mismatch_file():
-    testfile = str(get_test_data_path('version_mismatch.fits'))
-
-    with pytest.warns(AsdfConversionWarning, match="tag:stsci.edu:asdf/core/complex"):
-        with asdf.open(testfile,
-                ignore_version_mismatch=False) as fits_handle:
-            assert fits_handle.tree['a'] == complex(0j)
-
-    # Make sure warning does not occur when warning is ignored (default)
-    with assert_no_warnings(AsdfConversionWarning):
-        with asdf.open(testfile) as fits_handle:
-            assert fits_handle.tree['a'] == complex(0j)
-
-    with pytest.warns(AsdfConversionWarning, match="tag:stsci.edu:asdf/core/complex"):
-        with fits_embed.AsdfInFits.open(testfile,
-                ignore_version_mismatch=False) as fits_handle:
-            assert fits_handle.tree['a'] == complex(0j)
-
-    # Make sure warning does not occur when warning is ignored (default)
-    with assert_no_warnings(AsdfConversionWarning):
-        with fits_embed.AsdfInFits.open(testfile) as fits_handle:
-            assert fits_handle.tree['a'] == complex(0j)
 
 
 def test_serialize_table(tmpdir):